requirements.txt
# INSTRUCTIONS

# Follow the following instructions ON the raspberry shake!

# create requirements file with 
#   pip freeze -r ./requirements.txt

# download offline packages with
#   pip download -r ./requirements.txt -d ./offline_packages

# install packages with
#   pip install -r ./requirements.txt

## The following requirements were added by pip freeze:
pyserial==3.5
tomli==2.0.1
pymysql
pyserial